backend/foodgram/requirements.txt
Django==3.2
djoser
djangorestframework
